requirements.txt
docx2txt==0.8
textstat==0.7.3
python-docx==0.8.11
lxml==4.9.2
requests==2.22.0
bs4==0.0.1
beautifulsoup4==4.11.1
soupseive==2.3.2.post1
pdfminer.six==20221105
pycryptodome==3.16.0
pyphen==0.13.2
cryptography==39.0.0
pyOpenSSL==23.0.0
charset-normalizer==2.1.1
aiohttp==3.8.3
frozenlist==1.3.3
aiosignal==1.3.1
yarl==1.8.2
multidict==6.0.4
idna==2.8
attrs==19.3.0
async-timeout==4.0.2

